--- Content.Server/_ES/Masks/Blessed/Components/ESBlessedKillerMarkerComponent.cs
@@ -1,7 +1,0 @@
-namespace Content.Server._ES.Masks.Blessed.Components;
-
-/// <summary>
-///     Marks an entity which should die after a given time, as penance for killing the Blessed.
-/// </summary>
-[RegisterComponent]
-public sealed partial class ESBlessedKillerMarkerComponent : Component;
