+++ Content.Server/_ES/Masks/Martyr/Components/ESMartyrKillerMarkerComponent.cs
@@ -1,0 +1,7 @@
+namespace Content.Server._ES.Masks.Martyr.Components;
+
+/// <summary>
+///     Marks an entity which should die after a given time, as penance for killing the Martyr.
+/// </summary>
+[RegisterComponent]
+public sealed partial class ESMartyrKillerMarkerComponent : Component;
