+++ Content.Server/_ES/Masks/Martyr/ESMartyrSystem.cs
@@ -1,4 +1,4 @@
-using Content.Server._ES.Masks.Blessed.Components;
+using Content.Server._ES.Masks.Martyr.Components;
 using Content.Server._ES.Masks.Objectives;
 using Content.Server._ES.Masks.Objectives.Relays;
 using Content.Server.Administration;
@@ -8,18 +8,18 @@ using Content.Shared._ES.Masks;
 using Content.Shared.Body.Systems;
 using Robust.Shared.Player;
 using Robust.Shared.Prototypes;
-using Content.Shared._ES.Masks.Blessed;
+using Content.Shared._ES.Masks.Martyr;
 
-namespace Content.Server._ES.Masks.Blessed;
+namespace Content.Server._ES.Masks.Martyr;
 
 /// <summary>
-///     Handles gameplay logic for the Blessed mask--i.e., checking if they were killed by a crewmember,
+///     Handles gameplay logic for the Martyr mask--i.e., checking if they were killed by a crewmember,
 ///     and marking their killer to be killed later as a result.
 /// </summary>
-/// <seealso cref="ESBlessedComponent"/>
-/// <seealso cref="ESBlessedKillerMarkerComponent"/>
+/// <seealso cref="ESMartyrComponent"/>
+/// <seealso cref="ESMartyrKillerMarkerComponent"/>
 /// <seealso cref="ESBeKilledObjectiveSystem"/>
-public sealed class ESBlessedSystem : EntitySystem
+public sealed class ESMartyrSystem : EntitySystem
 {
     [Dependency] private readonly SuicideSystem _suicide = default!;
     [Dependency] private readonly SharedBodySystem _body = default!;
@@ -33,11 +33,11 @@ public sealed class ESBlessedSystem : EntitySystem
     {
         base.Initialize();
 
-        SubscribeLocalEvent<ESBlessedComponent, ESKillReportedEvent>(OnKillReported);
-        SubscribeLocalEvent<ESBlessedKillerMarkerComponent, ESBlessedKillerTimeToDieEvent>(OnTimeToDie);
+        SubscribeLocalEvent<ESMartyrComponent, ESKillReportedEvent>(OnKillReported);
+        SubscribeLocalEvent<ESMartyrKillerMarkerComponent, ESMartyrKillerTimeToDieEvent>(OnTimeToDie);
     }
 
-    private void OnTimeToDie(Entity<ESBlessedKillerMarkerComponent> ent, ref ESBlessedKillerTimeToDieEvent args)
+    private void OnTimeToDie(Entity<ESMartyrKillerMarkerComponent> ent, ref ESMartyrKillerTimeToDieEvent args)
     {
         if (!_suicide.Suicide(ent))
         {
@@ -49,7 +49,7 @@ public sealed class ESBlessedSystem : EntitySystem
     // we dont actually force this to be relayed --
     // instead we just assume that it will be relayed, if we are in the mind, because of our objective to be killed
     // if it isnt, then idk ur doing something wrong
-    private void OnKillReported(Entity<ESBlessedComponent> ent, ref ESKillReportedEvent args)
+    private void OnKillReported(Entity<ESMartyrComponent> ent, ref ESKillReportedEvent args)
     {
         if (!_beKilled.IsValidKill(args, KillerMustBeTroupe, out var killerMind))
             return;
@@ -57,14 +57,14 @@ public sealed class ESBlessedSystem : EntitySystem
         if (killerMind.Value.Comp.CurrentEntity is not { } killerBody)
             return;
 
-        EnsureComp<ESBlessedKillerMarkerComponent>(killerBody);
-        _timer.SpawnTimer(killerBody, ent.Comp.TimeBeforeKillerDeath, new ESBlessedKillerTimeToDieEvent());
+        EnsureComp<ESMartyrKillerMarkerComponent>(killerBody);
+        _timer.SpawnTimer(killerBody, ent.Comp.TimeBeforeKillerDeath, new ESMartyrKillerTimeToDieEvent());
 
         if (!TryComp<ActorComponent>(killerBody, out var actor))
             return;
 
-        var title = Loc.GetString("es-mask-blessed-killer-quickdialog-title");
-        var msg = Loc.GetString("es-mask-blessed-killer-quickdialog-msg");
+        var title = Loc.GetString("es-mask-martyr-killer-quickdialog-title");
+        var msg = Loc.GetString("es-mask-martyr-killer-quickdialog-msg");
 
         // we are kind of misusing quickdialogs by just using them as a persistent UI popup rather than
         // entering any data, so we just ignore it with an empty action
